refactor(identity): split the crate into focused modules; empty the size allowlist
The last file on the guardrail allowlist, split verbatim with no logic
changes:

  lib.rs      constants, IdentityError, re-exports (public API unchanged)
  device.rs   DeviceIdentity + stored-identity codec
  fs.rs       the private crash-safe dirfd write/read machinery
  roles.rs    the sealed signing roles
  signer.rs   TypedSigner + issuer/key-id/header helpers
  trust.rs    KeyValidity, RoleTrustStore, VerifiedCosePayload
  tests.rs    the unchanged suite

Cross-module internals are pub(crate); nothing new enters the public
surface. The guardrail allowlist is now empty — and stays that way: split
modules instead of adding entries. 163 tests pass unchanged.
